# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
  role: string;
}

export interface UserDto {
  id: string;
  name: string;
  email: string;
  role: string;
  siteId: string;
  status: string;
  phone: string;
  unitNo?: string;
  plates?: string[];
  businessName?: string;
  businessCategory?: string;
  businessDescription?: string;
  businessAddress?: string;
  latitude?: number;
  longitude?: number;
  createdAt: string;
}

export interface AuthResponse {
  user: UserDto;
  token: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: string;
  phone: string;
  siteId?: string;
  siteName?: string;
  siteAddress?: string;
  unitNo?: string;
  plates?: string[];
  businessName?: string;
  businessCategory?: string;
  businessDescription?: string;
  businessAddress?: string;
  latitude?: number;
  longitude?: number;
}

export interface RegisterResult {
  success: boolean;
  message: string;
  user?: UserDto;
  token?: string;
}

export interface UpdateUserRequest {
  name?: string;
  phone?: string;
  unitNo?: string;
  plates?: string[];
  businessName?: string;
  businessCategory?: string;
  businessDescription?: string;
  businessAddress?: string;
  latitude?: number;
  longitude?: number;
}

export interface SiteDto {
  id: string;
  name: string;
  address: string;
  adminId: string;
  createdAt: string;
}

export interface NotificationDto {
  id: string;
  type: string;
  title: string;
  message: string;
  fromUserId: string;
  fromName: string;
  toRoles?: string[];
  toUserIds?: string[];
  siteId: string;
  readBy: string[];
  createdAt: string;
}

export interface SendNotificationRequest {
  type: string;
  title: string;
  message: string;
  fromUserId: string;
  fromName: string;
  toRoles?: string[];
  toUserIds?: string[];
  siteId: string;
}

export interface PaymentDto {
  id: string;
  siteId: string;
  title: string;
  amount: number;
  dueDate: string;
  type: string;
  description?: string;
  createdAt: string;
}

export interface CreatePaymentRequest {
  siteId: string;
  title: string;
  amount: number;
  dueDate: string;
  type: string;
  description?: string;
}

export interface UserPaymentDto {
  id: string;
  paymentId: string;
  userId: string;
  siteId: string;
  status: string;
  paidAt?: string;
}

export interface MessageDto {
  id: string;
  chatId: string;
  fromId: string;
  fromName: string;
  content: string;
  createdAt: string;
}

export interface SendMessageRequest {
  chatId: string;
  toId?: string;
  content: string;
}

export interface PackageDto {
  id: string;
  siteId: string;
  recipientUserId: string;
  recipientName: string;
  senderInfo: string;
  description: string;
  status: string;
  receivedAt: string;
  deliveredAt?: string;
}

export interface AddPackageRequest {
  siteId: string;
  recipientUserId: string;
  recipientName: string;
  senderInfo: string;
  description: string;
}

export interface UpdatePackageStatusRequest {
  status: string;
}

export type GetUsersParams = {
siteId?: string;
};

export type GetMessagesParams = {
chatId?: string;
};