# lib/api-zod/src/generated/types/addPackageRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddPackageRequest {
  siteId: string;
  recipientUserId: string;
  recipientName: string;
  senderInfo: string;
  description: string;
}