# lib/api-zod/src/generated/types/createPaymentRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePaymentRequest {
  siteId: string;
  title: string;
  amount: number;
  dueDate: string;
  type: string;
  description?: string;
}