# lib/api-zod/src/generated/types/messageDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageDto {
  id: string;
  chatId: string;
  fromId: string;
  fromName: string;
  content: string;
  createdAt: string;
}