# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: string;
  phone: string;
  siteId?: string;
  siteName?: string;
  siteAddress?: string;
  unitNo?: string;
  plates?: string[];
  businessName?: string;
  businessCategory?: string;
  businessDescription?: string;
  businessAddress?: string;
  latitude?: number;
  longitude?: number;
}