# lib/api-zod/src/generated/types/registerResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserDto } from './userDto';

export interface RegisterResult {
  success: boolean;
  message: string;
  user?: UserDto;
  token?: string;
}