# lib/api-zod/src/generated/types/sendNotificationRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendNotificationRequest {
  type: string;
  title: string;
  message: string;
  fromUserId: string;
  fromName: string;
  toRoles?: string[];
  toUserIds?: string[];
  siteId: string;
}