# lib/api-zod/src/generated/types/userDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserDto {
  id: string;
  name: string;
  email: string;
  role: string;
  siteId: string;
  status: string;
  phone: string;
  unitNo?: string;
  plates?: string[];
  businessName?: string;
  businessCategory?: string;
  businessDescription?: string;
  businessAddress?: string;
  latitude?: number;
  longitude?: number;
  createdAt: string;
}