# lib/api-zod/src/generated/types/userPaymentDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserPaymentDto {
  id: string;
  paymentId: string;
  userId: string;
  siteId: string;
  status: string;
  paidAt?: string;
}